Move claripy.transition into claripy.utils.transition
