+++ claripy/utils/transition.py
